+++ Deviot.py
@@ -7,18 +7,27 @@ from __future__ import division
 from __future__ import unicode_literals
 
 from os import path
+from sublime import windows
 from sublime_plugin import EventListener
 
 from .commands import *
 from .platformio.update import Update
 from .beginning.pio_install import PioInstall
-from .libraries.tools import get_setting, save_setting, get_phantoms, del_phantom, set_deviot_syntax
-from .libraries.paths import getBoardsFileDataPath, getMainMenuPath
+from .libraries.tools import get_setting, save_setting, set_deviot_syntax
+from .libraries.paths import getBoardsFileDataPath, getMainMenuPath, getPluginPath
 from .libraries.preferences_bridge import PreferencesBridge
 from .libraries.project_check import ProjectCheck
 
 def plugin_loaded():
+    # Load or fix the right deviot syntax file 
+    for window in windows():
+        for view in window.views():
+            set_deviot_syntax(view)
+
+    # Install PlatformIO
     PioInstall()
+
+    # Search updates
     Update().check_update_async()
 
     menu_path = getMainMenuPath()
@@ -28,6 +37,14 @@ def plugin_loaded():
         from .libraries.top_menu import TopMenu
         TopMenu().make_menu_files()
         save_setting('compile_lang', False)
+
+    # check if the syntax file exist
+    deviot_syntax = getPluginPath()
+    syntax_path = path.join(deviot_syntax, 'deviot.sublime-syntax')
+
+    if(not path.exists(syntax_path)):
+        active_window().run_command('deviot_rebuild_syntax')
+
 
 class DeviotListener(EventListener):
     def on_load(self, view):
@@ -47,24 +64,3 @@ class DeviotListener(EventListener):
             serial_monitor = serial.serial_monitor_dict.get(port_id, None)
             serial_monitor.stop()
             del serial.serial_monitor_dict[port_id]
-
-    def on_modified(self, view):
-        """On modify file
-        
-        checks the phantoms in the current view and remove it
-        when it's neccesary
-        
-        Arguments:
-            view {obj} -- sublime text object
-        """
-        is_iot = ProjectCheck().is_iot()
-        phantoms = get_phantoms()
-
-        if(not len(phantoms) and not is_iot):
-            return
-
-        line, column = view.rowcol(view.sel()[0].begin())
-        pname = 'error' + str(line + 1)
-
-        if(pname in phantoms and view.file_name()):
-            del_phantom(pname)