+++ libraries/__init__.py
@@ -1,4 +1,4 @@
-VERSION = (2, 0, 1)
+VERSION = (2, 1, 0)
 __version__ = ".".join([str(s) for s in VERSION])
 
 __title__ = "Deviot"