+++ platformio/initialize.py
@@ -6,11 +6,12 @@ from __future__ import print_function
 from __future__ import division
 from __future__ import unicode_literals
 
-from sys import exit
+from os import path, remove
+from shutil import copyfile
 from ..libraries import __version__ as version
 from ..libraries.project_check import ProjectCheck
 from ..libraries.messages import MessageQueue
-from ..libraries.tools import reset_phantoms
+from ..libraries.tools import save_sysetting
 
 class Initialize(ProjectCheck):
     """
@@ -26,16 +27,12 @@ class Initialize(ProjectCheck):
     def __init__(self):
         super(Initialize, self).__init__()
 
-        reset_phantoms()
-
         messages = MessageQueue("_deviot_starting{0}", version)
         messages.start_print()
         
         self.dprint = messages.put
         self.derror = messages.print_once
         self.dstop = messages.stop_print
-
-        self.fix_src()
 
     def add_board(self):
         """New Board
@@ -77,3 +74,52 @@ class Initialize(ProjectCheck):
 
         thread = Thread(target=self.add_board)
         thread.start()
+
+    def after_complete(self):
+        """At complete
+        
+        This method will run functions after complete a compilation
+        or upload an sketch. You should only put here a fuction or
+        a method
+        """
+
+        # remove the cpp temporal file
+        self.del_cpp_temp()
+
+        # remove src_filter from platformio.ini
+        self.exclude_ino()
+
+        # remove src_dir flag from platformio.ini
+        self.remove_src()
+
+        # stop message queue
+        self.dstop()
+
+        # none last action
+        save_sysetting('last_action', None)
+
+    def make_cpp_temp(self):
+        """cpp file
+        
+        copy a ino file in a new cpp file
+        """
+        extension = self.get_file_extension()
+
+        if(extension == 'ino'):
+            file_path = self.get_file_path()
+            cpp_file = file_path.replace('.ino', '.cpp')
+            
+            copyfile(file_path, cpp_file)
+            self.exclude_ino(file_path)
+
+    def del_cpp_temp(self):
+        """remove cpp
+        
+        remove the cpp file
+        """
+        file_path = self.get_file_path()
+        ino_file = file_path.replace('.cpp', '.ino')
+        cpp_file = file_path.replace('.ino', '.cpp')
+
+        if(path.exists(cpp_file) and path.exists(ino_file)):
+            remove(cpp_file)